requirements.txt
PyMySql
baidu-aip
bs4